Fold skew detection into aggregate
environment_skew.py had a single caller. Moving the detector next to the
code that reports it puts the whole stale-image path in one place, and
its tests join the aggregate reporting tests they belong with.
