+++ tools/perf_smoke_test/test/test_aggregate_reseed.py
@@ -3,13 +3,15 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""GPU-free checks for aggregate's post-gate under-filled-bucket detector.
+"""GPU-free checks for aggregate's post-gate reporting decisions.
 
-The detector flags a task for targeted reseeding when this run produced a valid
-measurement but the matching baseline window for its exact runtime_contract_hash
-has fewer than MIN_BASELINE_SAMPLES samples. These tests drive aggregate.main()
-in flat-file mode (no git) and assert the reseed_tasks GITHUB_OUTPUT signal that
-the gate's reseed job consumes.
+Two concerns are covered. The under-filled-bucket detector flags a task for
+targeted reseeding when this run produced a valid measurement but the matching
+baseline window for its exact runtime_contract_hash has fewer than
+MIN_BASELINE_SAMPLES samples; those tests drive aggregate.main() in flat-file
+mode (no git) and assert the reseed_tasks GITHUB_OUTPUT signal the gate's reseed
+job consumes. The skew detector decides whether a crash is blamed on the change
+under test or on an out-of-date CI image.
 """
 
 from __future__ import annotations
@@ -178,7 +180,56 @@ def test_reseed_not_flagged_without_valid_measurement(tmp_path, monkeypatch) -> 
     assert "reseed_tasks" not in outputs
 
 
-_STALE_IMAGE_LOG = "ImportError: cannot import name 'SolverNotifyFlags' from 'newton.solvers'"
+# Verbatim from the 2026-07-28 staging run, where source pinning a newer Newton
+# met an image built before SolverNotifyFlags was replaced by ModelFlags.
+_STALE_IMAGE_LOG = """
+Traceback (most recent call last):
+  File "/workspace/isaaclab/source/isaaclab_newton/isaaclab_newton/physics/newton_manager.py", line 21, in <module>
+    from newton.solvers import SolverNotifyFlags
+ImportError: cannot import name 'SolverNotifyFlags' from 'newton.solvers'
+"""
+
+
+def test_missing_newton_symbol_is_reported_as_stale_image() -> None:
+    """The crash that broke the last qualification run must be recognized."""
+    skew = aggregate.detect_dependency_skew(_STALE_IMAGE_LOG)
+
+    assert skew is not None
+    assert skew.package == "newton"
+    assert skew.symbol == "SolverNotifyFlags"
+    assert "no `SolverNotifyFlags`" in skew.describe()
+
+
+def test_missing_image_module_is_reported_as_stale_image() -> None:
+    """A package the image should provide but does not is also image skew."""
+    skew = aggregate.detect_dependency_skew("ModuleNotFoundError: No module named 'newton.solvers.kamino'")
+
+    assert skew is not None
+    assert skew.package == "newton"
+    assert skew.symbol is None
+    assert "is not installed" in skew.describe()
+
+
+def test_missing_module_attribute_is_reported_as_stale_image() -> None:
+    """Warp exposing no such attribute means the installed Warp predates the pin."""
+    skew = aggregate.detect_dependency_skew("AttributeError: module 'warp' has no attribute 'sparse_matmul'")
+
+    assert skew is not None
+    assert skew.package == "warp"
+    assert skew.symbol == "sparse_matmul"
+
+
+def test_missing_isaaclab_symbol_is_not_excused() -> None:
+    """Isaac Lab source is mounted from the PR, so its own broken import is a real defect."""
+    log = "ImportError: cannot import name 'ArticulationCfg' from 'isaaclab.assets'"
+
+    assert aggregate.detect_dependency_skew(log) is None
+
+
+def test_clean_log_reports_no_skew() -> None:
+    """A run that never raised an import error is not skew."""
+    assert aggregate.detect_dependency_skew("Step Frametimes: 3.1 3.0 3.2") is None
+    assert aggregate.detect_dependency_skew(None) is None
 
 
 def test_crash_from_stale_image_does_not_fail_the_gate(tmp_path, monkeypatch) -> None:
